Integrated consumer and producer code, removed excess code
